+++ src/main/java/wooteco/subway/domain/fare/FareCalculator.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.fare;
 
 public class FareCalculator {
