+++ src/main/java/wooteco/subway/domain/section/Sections.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.section;
 
 import java.util.HashSet;
 import java.util.LinkedList;
