+++ src/main/java/wooteco/subway/service/dto/LineResponse.java
@@ -1,7 +1,7 @@
 package wooteco.subway.service.dto;
 
 import java.util.List;
-import wooteco.subway.domain.Line;
+import wooteco.subway.domain.line.Line;
 
 public class LineResponse {
