+++ src/main/java/wooteco/subway/service/dto/StationResponse.java
@@ -1,6 +1,6 @@
 package wooteco.subway.service.dto;
 
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 public class StationResponse {
