+++ src/main/java/wooteco/subway/ui/dto/SectionRequest.java
@@ -1,6 +1,6 @@
 package wooteco.subway.ui.dto;
 
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.section.Section;
 
 public class SectionRequest {
