+++ src/test/java/wooteco/subway/domain/PathStrategyTest.java
@@ -5,6 +5,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.path.PathStrategy;
+import wooteco.subway.domain.path.ShortestPathStrategy;
+import wooteco.subway.domain.section.Section;
+import wooteco.subway.domain.station.Station;
 
 class PathStrategyTest {
