+++ src/test/java/wooteco/subway/domain/SectionsTest.java
@@ -12,6 +12,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import wooteco.subway.domain.section.Section;
+import wooteco.subway.domain.section.Sections;
 
 class SectionsTest {
